Shifting to Esp32WiFiManager and enable LCC
This is the first of a few steps towards further improving the stability of the ESP32CS code. In this step the LCC interface is now always enabled (WiFi side only) and the WiFi connection management has been shifted to the Esp32WiFiManager.

For now the CS Static IP support is dropped but it will return soon.

The InfoScreen is currently not displaying the CS IP address when using an OLED display, this is to be fixed soon.