:hammer: Transform HC12 code to proper StateFlow
The HC12 adapter now uses read_single/write_repeated with ::select() with the UART device. If the /dev/uart/<num> fails to open the flow will exit automatically with a log message but keep the CS up and running. This can be improved to report the failure via the config portal in a subsequent commit.