:ambulance: Split update into update/update_bus.
This is needed to ensure that NeoPixelBus has finished the previous Show() call prior to calling Show() again. If the previous call is still executing the update_bus method will go to try again on next call to send the current state. This may cause LED updates to be delayed slightly but prevents occasional corruption of LED color data.