:construction: Disabled old HTTP code
The new Httpd code is mostly stable and the work has been started to shift over the existing endpoints to the new API. More work needs to be done still.

OTA is failing in Chrome some of the time, likely due to corruption in the data stream due to parsing failures. This will need more testing before release.