:tractor: :construction: Restructure of code and move to ESP-IDF v4.
All arduino-esp32 dependencies are being removed. As such some functionality is 100% broken currently (status display, nextion, loconet, etc) but basic functionality appears to be intact. LCC CAN is not wired up yet as the CAN driver needs to be ported still.

Building from code will require:
1) ESP-IDF (https://docs.espressif.com/projects/esp-idf/en/release-v4.0/get-started/index.html)
2) GZIP (must be in path)
3) Run IDF_PATH/tools/idf.py menuconfig to configure CS features/functionality. The include files will not be used now!

Before flashing to the ESP32 it is recommended to run "idf.py erase_flash" to start with a fresh configuation as some options have changed.

Flashing the code to the ESP32 should be done with: idf.py flash monitor
This will start the serial monitor immediately after flashing. It is recommended to have the toolchain bin directory in the PATH when doing this so that stacktraces are automatically decoded.