:tractor: Splitting up RMTTrackDevice. :pencil: Adding validation for config json. :pencil: remove CAN RX/TX pins from config.
The existing RMTTrackDevice class was becoming too large and complicated to manage easily. The addition of RailCom was a big contributor to the complications.

The new code now creates two instances of RMTTrackDevice (OPS and PROG) which will operate independently (RMT TX complete ISR is not independent though). RailCom functionality is shifting to RailcomDriver instead which will be for OPS only.