:ambulance: Fixing DCC signal generation
With the refactoring/split of RMTTrackDevice recently the ENCODE_PACKET() macro was causing the RMT to only transmit one packet. This caused the signal pin to remain in a LOW state indefinitely.

Reintroduced the encode_next_packet() method with the ENCODE_PACKET() macro expanded in-place in the method with a TODO marker for marklin-motorola encoding to be added in the future.

An unexpected bug was found with using uninitialized<type>, it appears that it does not correctly initialize all members of the underlying data type when emplace() is called. This was leading to packetQueue_ being null when the RMT TX complete ISR was fired.