:pencil: Remove RailCom short pin definition.
This pin is not actively used by the ESP32 and likely doesn't need to be used directly as it should be disabling the FETs as part of the schematic which will cause the RailCom reads to fail (or otherwise be corrupted) which will cause them to be discarded. The short detection outside of the cutout will detect and disable track output if the short persists.